% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateFrequency.R
\name{calculate.frequencies}
\alias{calculate.frequencies}
\title{Calculates clone size frequencies if clone counts are given as input}
\usage{
calculate.frequencies(df)
}
\arguments{
\item{clones}{A dataframe with a column counts that contains the counts for
each mutation in a clone \code{clones}}
}
\value{
clones Returns vector with frequencies
}
\description{
Calculates clone size frequencies if clone counts are given as input
}
\examples{
clones <- calculate.frequencies(clones)
}
\keyword{clone}
\keyword{frequency}

