% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateCoordinates.R
\name{calculate.gamma.coordinates}
\alias{calculate.gamma.coordinates}
\title{Calculate gamma coordinates to draw the polygons}
\usage{
calculate.gamma.coordinates(freq_clones, xaxis_length, yaxis_length,
  xy_start = c(0, 0), debug = FALSE)
}
\arguments{
\item{freq_clones}{A vector with the frequency of the clones \code{freq_clones}}

\item{xaxis_length}{Length of the x-axis \code{xaxis_length}}

\item{yaxis_length}{Length of the y-axis \code{yaxis_length}}

\item{xy_start}{Coordinates for the starting point of the plot \code{xy_start}}
}
\value{
dataframe Returns a dataframe with x and y coordinates for point gamma
}
\description{
(beta)
             ^
           + +
 (gamma) < + +
           + +
             v
             (alpha)
}
\examples{
alpha_beta <- calculate.alpha.beta.coordinates(df[df$clone==1, ],
                                              xaxis_length=4,
                                              yaxis_length=yaxis_length_per_clone[1],
                                              xy_start=c(0,0))
}
\keyword{alpha}
\keyword{beta}
\keyword{coordinates}
\keyword{gamma}

