% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateLengthXaxis.R
\name{calculate.xaxis.length.clones}
\alias{calculate.xaxis.length.clones}
\title{Calculate the length of the x-axis}
\usage{
calculate.xaxis.length.clones(freq_clones, xaxis_length = 4)
}
\arguments{
\item{freq_clones}{Frequency of clones \code{freq_clones}}

\item{xaxis_length}{Length of the xaxis \code{xaxis_length}}
}
\value{
xaxis_length_per_clone
}
\description{
Calculate the length of the x-axis
}
\examples{
yaxis_length_per_clone_1 <- calculate.xaxis.length.clones(freq_clones_1, xaxis_length=4)
}
\keyword{clone}
\keyword{frequency}

