% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calculateLengthYaxis.R
\name{calculate.yaxis.length.clones}
\alias{calculate.yaxis.length.clones}
\title{Calculate the length of the y-axis}
\usage{
calculate.yaxis.length.clones(freq_clones, yaxis_length = 4, debug = FALSE)
}
\arguments{
\item{freq_clones}{Frequency of clones \code{freq_clones}}

\item{yaxis_length}{Length of the yaxis \code{yaxis_length}}
}
\value{
yaxis_length_per_clone Returns a vector
}
\description{
Calculate the length of the y-axis
}
\examples{
yaxis_length_per_clone <- calculate.yaxis.length.clones(df_freq, yaxis_length=4, debug=TRUE)
}
\keyword{clone}
\keyword{frequency}

