% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/drawPolygon.R
\name{draw.clonal.evolution.plot}
\alias{draw.clonal.evolution.plot}
\title{Draw the polygone given as input the coordinates}
\usage{
draw.clonal.evolution.plot(d_plot)
}
\arguments{
\item{d}{dataframe with coordinates \code{d}}
}
\value{
g_obj a ggplot object ready to plot with plot(g_obj)
}
\description{
Draw the polygone given as input the coordinates
}
\examples{
g_plot <- draw.clonal.evolution.plot(d)
}
\keyword{plot}
\keyword{polygon}

