% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getCoordinates.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Helper function that calls other functions to calculate coordinates}
\usage{
get_coordinates(df, yaxis_length = 4, debug = TRUE)
}
\arguments{
\item{df}{dataframe with calculated frequencies \code{df}}

\item{yaxis_length}{Length of the yaxis [num] \code{yaxis_length}}

\item{debug}{TRUE gives verbose output [boolean] \code{debug}}
}
\value{
dataframe with x & y coordinates for all polygons
}
\description{
Helper function that calls other functions to calculate coordinates
}
\examples{
yaxis_length_per_clone <- calculate.yaxis.length.clones(df_freq, yaxis_length=4, debug=TRUE)
}
\keyword{coordinates}

